fix: self-heal Windows UAC on first CLI invocation
`bun add -g` skips lifecycle scripts unconditionally, so the postinstall
script added in the previous fix never ran for bun-installed users —
they kept seeing UAC at startup despite the patch shipping.

Extract the manifest-writing logic into src/windows-manifest.ts and call
ensureWindowsManifest() at the top of cli.ts. The first launch under bun
still triggers UAC (Windows evaluates the binary before any user-space
code runs), but the running process drops the external `.exe.manifest`
and bumps the shim's mtime to invalidate Windows' cached elevation
decision. Every subsequent launch is silent.

The postinstall script now delegates to the same module, keeping a
single source of truth for the manifest payload and shim discovery.
