+++ src/cli.ts
@@ -7,6 +7,13 @@
  * the process with code 1.
  */
 import { main } from "./index.ts";
+import { ensureWindowsManifest } from "./windows-manifest.ts";
+
+// Drop external asInvoker manifest next to the bun-generated `.exe` shim on
+// Windows so subsequent launches do not trigger UAC. First launch under bun
+// still UACs (the OS evaluates the binary before any code runs); from then on
+// it is silent. No-op on non-Windows. See windows-manifest.ts for context.
+ensureWindowsManifest();
 
 main().catch((err) => {
   if (err instanceof Error) {
